requirements.txt
beautifulsoup4
pillow
requests